Modify the compile problem in win

+++ exporter/src/utils/PsdTextAttribute.h
@@ -17,6 +17,7 @@
 /////////////////////////////////////////////////////////////////////////////////////////////////
 
 #pragma once
+#include <memory>
 #include <string>
 #include <vector>
 
@@ -51,7 +52,7 @@ class PsdTextAttribute {
   void buildAttributeArray(int startPos, int endPos);
 
  private:
-  int stringFormatTransform(std::vector<char>& dst, const uint8_t* src, int len);
+  size_t stringFormatTransform(std::vector<char>& dst, const uint8_t* src, int len);
 
   std::vector<char> mem = {};
   char* src = nullptr;
